Finish Exercise 13.12

+++ controllers/blogs.js
@@ -1,15 +1,27 @@
 const router = require("express").Router();
-const { Blog } = require("../models");
+const { Blog, User } = require("../models");
+const { tokenExtractor } = require("../util/middleware");
 
 router.get("/", async (req, res) => {
-  const blogs = await Blog.findAll();
+  const blogs = await Blog.findAll({
+    attributes: { exclude: ["userId"] },
+    include: {
+      model: User,
+      attributes: ["name"],
+    },
+  });
   console.log(JSON.stringify(blogs, null, 2));
   res.json(blogs);
 });
 
-router.post("/", async (req, res) => {
+router.post("/", tokenExtractor, async (req, res) => {
   try {
-    const blog = await Blog.create(req.body);
+    const user = await User.findByPk(req.decodedToken.id);
+    const blog = await Blog.create({
+      ...req.body,
+      userId: user.id,
+      date: new Date(),
+    });
     res.json(blog);
   } catch (error) {
     return res.status(400).json({ error });
@@ -48,13 +60,19 @@ router.put("/:id", blogFinder, async (req, res) => {
   }
 });
 
-router.delete("/:id", blogFinder, async (req, res) => {
+router.delete("/:id", tokenExtractor, blogFinder, async (req, res, next) => {
   try {
-    if (req.blog) {
-      await Blog.destroy({ where: { id: req.params.id } });
-      res.status(204);
-    } else {
-      throw new Error("blog not found");
+    if (req.decodedToken) {
+      if (req.blog) {
+        if (req.blog.userId === req.decodedToken.id) {
+          await Blog.destroy({ where: { id: req.params.id } });
+          res.status(204);
+        } else {
+          return res.status(401).json({ error: "unauthorized" });
+        }
+      } else {
+        throw new Error("blog not found");
+      }
     }
   } catch (error) {
     next(error);
